requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
websockets==14.1

# Google Cloud Services
google-generativeai==0.8.4
google-cloud-texttospeech==2.21.1
google-cloud-firestore==2.19.0
google-cloud-logging==3.11.3
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-monitoring==2.24.1

# Security & Utilities
python-multipart==0.0.20
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Code Quality
ruff==0.8.6
